refactor utils in tests
