+++ tests/backend/test_keras_backend.py
@@ -9,148 +9,29 @@ import numpy as np
 import pytest
 import six
 
-from fuel.datasets.hdf5 import H5PYDataset
-from fuel.schemes import SequentialScheme
-from fuel.streams import DataStream
-from fuel.transformers import ScaleAndShift
-from keras.layers import Dense
-from keras.layers import Dropout
-from keras.layers import Input
-from keras.models import Model
-from keras.models import Sequential
-from keras.utils import np_utils
-from keras.utils.test_utils import get_test_data
-
 import alp.appcom.utils as utls
 from alp.appcom.core import Experiment
 from alp.appcom.utils import switch_backend
-from alp.appcom.utils import to_fuel_h5
 from alp.backend import keras_backend as KTB
 from alp.backend.common import open_dataset_gen
 from alp.backend.keras_backend import get_function_name
 from alp.backend.keras_backend import model_from_dict_w_opt
 from alp.backend.keras_backend import serialize
 from alp.backend.keras_backend import to_dict_w_opt
+from alp.utils.utils_tests import batch_size
+from alp.utils.utils_tests import close_gens
+from alp.utils.utils_tests import input_dim
+from alp.utils.utils_tests import make_data
+from alp.utils.utils_tests import make_gen
+from alp.utils.utils_tests import model
+from alp.utils.utils_tests import return_custom
+from alp.utils.utils_tests import sequential
+from alp.utils.utils_tests import test_samples
+from alp.utils.utils_tests import train_samples
 
 
-input_dim = 2
-nb_hidden = 4
-nb_class = 2
-batch_size = 4
-train_samples = 256
-test_samples = 128
 NAME = keras.__name__
 VERSION = keras.__version__
-
-
-def close_gens(gen, data, data_stream):
-    gen.close()
-    data.close(None)
-    data_stream.close()
-
-
-def make_data():
-    (X_tr, y_tr), (X_te, y_te) = get_test_data(nb_train=train_samples,
-                                               nb_test=test_samples,
-                                               input_shape=(input_dim,),
-                                               classification=True,
-                                               nb_class=nb_class)
-
-    y_tr = np_utils.to_categorical(y_tr)
-    y_te = np_utils.to_categorical(y_te)
-
-    data, data_val = dict(), dict()
-
-    data["X"] = X_tr
-    data["y"] = y_tr
-
-    data_val["X"] = X_te
-    data_val["y"] = y_te
-    return data, data_val
-
-
-def dump_data():
-    data, data_val = make_data()
-    inputs = [np.concatenate([data['X'], data_val['X']])]
-    outputs = [np.concatenate([data['y'], data_val['y']])]
-
-    file_name = 'test_data'
-    scale = 1.0 / inputs[0].std(axis=0)
-    shift = - scale * inputs[0].mean(axis=0)
-
-    file_path, i_names, o_names = to_fuel_h5(inputs, outputs, [0, 256],
-                                             ['train', 'test'],
-                                             file_name,
-                                             '/data_generator')
-    return file_path, scale, shift, i_names, o_names
-
-file_path, scale, shift, i_names, o_names = dump_data()
-
-
-def make_gen(examples=4):
-    file_path_f = file_path
-    names_select = i_names
-    train_set = H5PYDataset(file_path_f,
-                            which_sets=('train', 'test'))
-
-    scheme = SequentialScheme(examples=examples, batch_size=batch_size)
-
-    data_stream_train = DataStream(dataset=train_set, iteration_scheme=scheme)
-
-    stand_stream_train = ScaleAndShift(data_stream=data_stream_train,
-                                       scale=scale, shift=shift,
-                                       which_sources=(names_select[-1],))
-    return stand_stream_train, train_set, data_stream_train
-
-
-def return_custom():
-    import keras.backend as K
-    from keras.engine import Layer
-    class Dropout_cust(Layer):
-        '''Applies Dropout to the input.
-        '''
-
-        def __init__(self, p, **kwargs):
-            self.p = p
-            if 0. < self.p < 1.:
-                self.uses_learning_phase = True
-            self.supports_masking = True
-            super(Dropout_cust, self).__init__(**kwargs)
-
-        def call(self, x, mask=None):
-            if 0. < self.p < 1.:
-                x = K.in_train_phase(K.dropout(x, level=self.p), x)
-            return x
-
-        def get_config(self):
-            config = {'p': self.p}
-            base_config = super(Dropout_cust, self).get_config()
-            return dict(list(base_config.items()) + list(config.items()))
-
-    return Dropout_cust
-
-
-def sequential(custom=False):
-    model = Sequential()
-    model.add(Dense(nb_hidden, input_dim=input_dim, activation='relu'))
-    model.add(Dense(nb_class, activation='softmax'))
-    model.add(Dropout(0.5))
-    if custom:
-        model.add(return_custom()(0.5))
-    return model
-
-
-def model(custom=False):
-    inputs = Input(shape=(input_dim,), name='X')
-
-    x = Dense(nb_hidden, activation='relu')(inputs)
-    x = Dense(nb_hidden, activation='relu')(x)
-    predictions = Dense(nb_class,
-                        activation='softmax',
-                        name='main_loss')(x)
-
-    model = Model(input=inputs, output=predictions)
-    return model
 
 
 def prepare_model(get_model, get_loss_metric, custom):
@@ -200,23 +81,23 @@ def prepare_model(get_model, get_loss_metric, custom):
 def get_generators(request):
     if request.param == 'one to many':
         nb_train, nb_val = 4, 48
-        gen_t, data_t, d_stream_t = make_gen(nb_train)
-        gen, data, d_stream = make_gen(nb_val)
+        gen_t, data_t, d_stream_t = make_gen(batch_size, nb_train)
+        gen, data, d_stream = make_gen(batch_size, nb_val)
 
     elif request.param == 'many to one':
         nb_train, nb_val = 48, 4
-        gen_t, data_t, d_stream_t = make_gen(nb_train)
-        gen, data, d_stream = make_gen(nb_val)
+        gen_t, data_t, d_stream_t = make_gen(batch_size, nb_train)
+        gen, data, d_stream = make_gen(batch_size, nb_val)
 
     elif request.param == 'val 1':
         nb_train, nb_val = 4, 4
-        gen_t, data_t, d_stream_t = make_gen(nb_train)
-        gen, data, d_stream = make_gen(nb_val)
+        gen_t, data_t, d_stream_t = make_gen(batch_size, nb_train)
+        gen, data, d_stream = make_gen(batch_size, nb_val)
 
     elif request.param == 'many to many':
         nb_train, nb_val = 48, 48
-        gen_t, data_t, d_stream_t = make_gen(nb_train)
-        gen, data, d_stream = make_gen(nb_val)
+        gen_t, data_t, d_stream_t = make_gen(batch_size, nb_train)
+        gen, data, d_stream = make_gen(batch_size, nb_val)
 
     return gen_t, data_t, d_stream_t, gen, data, d_stream, (nb_train, nb_val)
 
@@ -292,7 +173,7 @@ class TestExperiment:
 
     def test_experiment_fit(self, get_model, get_loss_metric,
                             get_custom_l):
-        data, data_val = make_data()
+        data, data_val = make_data(train_samples, test_samples)
         model, metrics, cust_objects = prepare_model(get_model(get_custom_l),
                                                      get_loss_metric,
                                                      get_custom_l)
@@ -320,7 +201,7 @@ class TestExperiment:
 
     def test_experiment_fit_async(self, get_model, get_loss_metric,
                                   get_custom_l):
-        data, data_val = make_data()
+        data, data_val = make_data(train_samples, test_samples)
         model, metrics, cust_objects = prepare_model(get_model(get_custom_l),
                                                      get_loss_metric,
                                                      get_custom_l)
@@ -368,13 +249,13 @@ class TestExperiment:
                                                      get_custom_l)
 
         model_name = model.__class__.__name__
-        _, data_val_use = make_data()
+        _, data_val_use = make_data(train_samples, test_samples)
         expe = Experiment(model)
 
         for val in [1, data_val_use]:
-            gen, data, data_stream = make_gen()
+            gen, data, data_stream = make_gen(batch_size)
             if val == 1:
-                val, data_2, data_stream_2 = make_gen()
+                val, data_2, data_stream_2 = make_gen(batch_size)
             expe.fit_gen([gen], [val], nb_epoch=2,
                          model=model,
                          metrics=metrics,
@@ -398,14 +279,14 @@ class TestExperiment:
                                                      get_loss_metric,
                                                      get_custom_l)
 
-        _, data_val_use = make_data()
+        _, data_val_use = make_data(train_samples, test_samples)
         expe = Experiment(model)
 
         expected_value = 2
         for val in [None, 1, data_val_use]:
-            gen, data, data_stream = make_gen()
+            gen, data, data_stream = make_gen(batch_size)
             if val == 1:
-                val, data_2, data_stream_2 = make_gen()
+                val, data_2, data_stream_2 = make_gen(batch_size)
             _, thread = expe.fit_gen_async([gen], [val], nb_epoch=2,
                                            model=model,
                                            metrics=metrics,
@@ -457,7 +338,7 @@ class TestExperiment:
                                                      False)
 
         model_name = model.__class__.__name__
-        data, data_val = make_data()
+        data, data_val = make_data(train_samples, test_samples)
 
         expe = Experiment(model)
         expe.fit([data], [data_val], nb_epoch=2,
@@ -501,7 +382,7 @@ class TestBackendFunctions:
 
     def test_fit(self, get_model):
         "Test the training of a serialized model"
-        data, data_val = make_data()
+        data, data_val = make_data(train_samples, test_samples)
 
         model = get_model()
         model.compile(loss='categorical_crossentropy',
@@ -525,7 +406,7 @@ class TestBackendFunctions:
 
     def test_predict(self, get_model):
         """Test to predict using the backend"""
-        data, data_val = make_data()
+        data, data_val = make_data(train_samples, test_samples)
         model = get_model()
         model.compile(optimizer='sgd', loss='categorical_crossentropy')
 
@@ -562,7 +443,7 @@ def test_utils():
     assert get_function_name("bob") == "bob"
     test_switch = switch_backend('sklearn')
     assert test_switch is not None
-    gen, data, data_stream = make_gen()
+    gen, data, data_stream = make_gen(batch_size)
     open_dataset_gen(data_stream)
     gen.close()
     data.close(None)
